Move publish in run directory
