+++ test/unit/lib/services/slack/commands_test.js
@@ -4,15 +4,16 @@ const { sinon } = require('../../../../test-helper');
 const {
   createAndDeployPixDatawarehouse,
 } = require('../../../../../common/services/slack/commands');
-const releasesServices = require('../../../../../lib/services/releases');
+const releasesServicesFromBuild = require('../../../../../build/services/releases');
+const releasesServicesFromRun = require('../../../../../run/services/releases');
 const githubServices = require('../../../../../common/services/github');
 
 describe('Services | Slack | Commands', () => {
   beforeEach(() => {
     // given
     sinon.stub(axios, 'post');
-    sinon.stub(releasesServices, 'publishPixRepo').resolves();
-    sinon.stub(releasesServices, 'deployPixRepo').resolves();
+    sinon.stub(releasesServicesFromBuild, 'publishPixRepo').resolves();
+    sinon.stub(releasesServicesFromRun, 'deployPixRepo').resolves();
     sinon.stub(githubServices, 'getLatestReleaseTag').resolves('v1.0.0');
   });
 
@@ -26,7 +27,7 @@ describe('Services | Slack | Commands', () => {
 
     it('should publish a new release', () => {
       // then
-      sinon.assert.calledWith(releasesServices.publishPixRepo, 'pix-db-replication', 'minor');
+      sinon.assert.calledWith(releasesServicesFromBuild.publishPixRepo, 'pix-db-replication', 'minor');
     });
 
     it('should retrieve the last release tag from GitHub', () => {
@@ -36,7 +37,7 @@ describe('Services | Slack | Commands', () => {
 
     it('should deploy the release', () => {
       // then
-      sinon.assert.calledWith(releasesServices.deployPixRepo);
+      sinon.assert.calledWith(releasesServicesFromRun.deployPixRepo);
     });
   });
 });
